pyproject.toml
[project]
name = "networksecurity"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "certifi>=2025.11.12",
    "dill>=0.4.0",
    "dotenv>=0.9.9",
    "numpy>=2.3.5",
    "pandas>=2.3.3",
    "pyaml>=25.7.0",
    "pymongo>=4.15.4",
    "scikit-learn>=1.7.2",
]
